+++ guillotina/security/__init__.py
@@ -1,0 +1,4 @@
+from . import cache
+
+
+security_map_cache = cache.SecurityMapCacheManager()